Track simple stats. G300 prints current values, if S0 is included, the values are reset. At present the only tracked stat is the count of calls to begin move that are made outside the interrupt context. Once a print is in progress, every call to begin move outside of the interrupt context is an undesirable pause.

+++ demitasse/gcode_parse.h
@@ -134,7 +134,7 @@ typedef struct {
 	};
 
 	uint8_t						G;				///< G command number
-	uint8_t						M;				///< M command number
+	uint16_t					M;				///< M command number
 	TARGET						target;		///< target position: X, Y, Z, E and F
 
 	int16_t						S;				///< S word (various uses)
